+++ engine/src/main/java/org/camunda/bpm/engine/runtime/Job.java
@@ -97,5 +97,5 @@ public interface Job {
    *
    * @since 7.4
    */
-  int getPriority();
+  long getPriority();
 }
