+++ engine/src/test/java/org/camunda/bpm/engine/test/api/runtime/TestOrderingUtil.java
@@ -430,7 +430,7 @@ public class TestOrderingUtil {
   public static NullTolerantComparator<HistoricJobLog> historicJobLogByJobPriority() {
     return new NullTolerantComparator<HistoricJobLog>() {
       public int compare(HistoricJobLog o1, HistoricJobLog o2) {
-        return Integer.valueOf(o1.getJobPriority()).compareTo(Integer.valueOf(o2.getJobPriority()));
+        return Long.valueOf(o1.getJobPriority()).compareTo(Long.valueOf(o2.getJobPriority()));
       }
 
       public boolean hasNullProperty(HistoricJobLog object) {
@@ -460,8 +460,8 @@ public class TestOrderingUtil {
 
       @Override
       public int compare(Job o1, Job o2) {
-        return Integer.valueOf(o1.getPriority())
-            .compareTo(Integer.valueOf(o2.getPriority()));
+        return Long.valueOf(o1.getPriority())
+            .compareTo(Long.valueOf(o2.getPriority()));
       }
 
       @Override
